basic CURD working for consultants

+++ DigitalentCoreApp/Controllers/ConsultantsController.cs
@@ -19,6 +19,15 @@ namespace DigitalentCoreApp.Controllers
         }
 
         // GET: Consultants
+    //        synchronous programming is the default mode for ASP.NET Core and EF Core.
+    //        A web server has a limited number of threads available, and in high load situations all of the available threads might be in use.When that 
+    //        happens, the server can't process new requests until the threads are freed up. With synchronous code, many threads may be tied 
+    //        up while they aren't actually doing any work because they're waiting for I/O to complete. With asynchronous code, when a process 
+    //        is waiting for I/O to complete, its thread is freed up for the server to use for processing other requests. As a result, asynchronous
+    //        code enables server resources to be used more efficiently, and the server is enabled to handle more traffic without delays.
+    //        Asynchronous code does introduce a small amount of overhead at run time, but for low traffic situations the performance hit is negligible, 
+    //        while for high traffic situations, the potential performance improvement is substantial.
+    //      In the following code, the async keyword, Task<T> return value, await keyword, and ToListAsync method make the code execute asynchronously.
         public async Task<IActionResult> Index()
         {
             return View(await _context.Consultants.ToListAsync());
@@ -33,7 +42,11 @@ namespace DigitalentCoreApp.Controllers
             }
 
             var consultant = await _context.Consultants
+                .Include(c => c.Assignments)
+                .ThenInclude(a => a.Project)
+                .AsNoTracking()
                 .SingleOrDefaultAsync(m => m.ID == id);
+  
             if (consultant == null)
             {
                 return NotFound();
@@ -53,13 +66,23 @@ namespace DigitalentCoreApp.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,FirstName,LastName,BirthDate,Phone,Adress,ZipCode")] Consultant consultant)
+        public async Task<IActionResult> Create([Bind("FirstName,LastName,BirthDate,Phone,Adress,ZipCode")] Consultant consultant)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(consultant);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(consultant);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.
+                ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the problem persists " +
+                    "see your system administrator.");
             }
             return View(consultant);
         }
@@ -81,42 +104,43 @@ namespace DigitalentCoreApp.Controllers
         }
 
         // POST: Consultants/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        // implement a security best practice to prevent overposting. No bind
+        
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,FirstName,LastName,BirthDate,Phone,Adress,ZipCode")] Consultant consultant)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != consultant.ID)
+            if (id == null)
             {
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            var consultantToUpdate = await _context.Consultants.SingleOrDefaultAsync(c => c.ID == id);
+            if (await TryUpdateModelAsync<Consultant>(
+                consultantToUpdate,
+                "",
+                c => c.FirstName, c => c.LastName, c => c.BirthDate, c => c.Phone, c => c.Adress, c => c.ZipCode))
             {
                 try
                 {
-                    _context.Update(consultant);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException /* ex */)
                 {
-                    if (!ConsultantExists(consultant.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    //Log the error (uncomment ex variable name and write a log.)
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                        "Try again, and if the problem persists, " +
+                        "see your system administrator.");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(consultant);
+
+            return View(consultantToUpdate);
         }
 
         // GET: Consultants/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        // an optional parameter that indicates whether the method was called after a failure to save changes.
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -124,24 +148,47 @@ namespace DigitalentCoreApp.Controllers
             }
 
             var consultant = await _context.Consultants
+                .AsNoTracking()
                 .SingleOrDefaultAsync(m => m.ID == id);
             if (consultant == null)
             {
                 return NotFound();
             }
 
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
+            }
+
             return View(consultant);
         }
 
         // POST: Consultants/Delete/5
+
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var consultant = await _context.Consultants.SingleOrDefaultAsync(m => m.ID == id);
-            _context.Consultants.Remove(consultant);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var consultant = await _context.Consultants
+                .AsNoTracking()
+                .SingleOrDefaultAsync(m => m.ID == id);
+            if (consultant == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+                _context.Consultants.Remove(consultant);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool ConsultantExists(int id)
